backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "browser-use>=0.2.0",
    "uagents>=0.10.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "twilio>=9.10.4",
    "langchain-anthropic>=1.4.0",
    "stripe>=15.0.1",
    "cosmpy>=0.11.2",
    "httpx>=0.27.0",
    "requests>=2.33.1",
]
